Fix docker registry storage (minio)

+++ scripts/run_background_services.sh
@@ -20,28 +20,57 @@ if [ -d /hosthome ]; then
     fi
 fi
 
+##################
+# Get master URL #
+##################
+if [ -z $MASTER_ADDRESS ]; then
+    read -p "Advertise URL: " MASTER_ADDRESS
+fi
+if [ -z $MINIO_PORT ]; then
+    read -p "MinIO Port (9000): " MINIO_PORT
+    if [ -z $MINIO_PORT ]; then
+        MINIO_PORT=9000
+    fi
+fi
+if [ -z $REGISTRY_PORT ]; then
+    read -p "Registry Port (5000): " REGISTRY_PORT
+    if [ -z $REGISTRY_PORT ]; then
+        REGISTRY_PORT=5000
+    fi
+fi
+
+# Access and secret key
+if [ -z $ACCESS_KEY ]; then
+    read -p "MinIO Access Key (MLAPPDEPLOY): " ACCESS_KEY
+    if [ -z $ACCESS_KEY ]; then
+        ACCESS_KEY=MLAPPDEPLOY
+    elif [ ${#ACCESS_KEY} -lt 3 ]; then
+        echo "Access key length should be between minimum 3 characters in length."
+        exit 1
+    fi
+fi
+if [ -z $SECRET_KEY ]; then
+    read -p "MinIO Secret Key (MLAPPDEPLOY): " SECRET_KEY
+    if [ -z $SECRET_KEY ]; then
+        SECRET_KEY=MLAPPDEPLOY
+    elif [ ${#SECRET_KEY} -lt 8 ]; then
+        echo "Secret key should be in between 8 and 40 characters."
+        exit 1
+    fi
+fi
 
 ################
 # MinIO Server #
 ################
-
 MINIO_DATA=$MLAD/master/data
 MINIO_CONFIG=$MLAD/master/minio
 
 mkdir -p $MINIO_DATA $MINIO_CONFIG
 
-# Access and secret key
-if [ -z $ACCESS_KEY ]; then
-    ACCESS_KEY=MLAPPDEPLOY
-fi
-if [ -z $SECRET_KEY ]; then
-    SECRET_KEY=MLAPPDEPLOY
-fi
-
 # Run minio server
 docker stop minio > /dev/null 2>&1
 docker rm minio > /dev/null 2>&1
-docker run -d -p 9000:9000 --restart=always --name minio -e "MINIO_ACCESS_KEY=$ACCESS_KEY" -e "MINIO_SECRET_KEY=$SECRET_KEY" -v $MINIO_DATA:/data -v $MINIO_CONFIG:/root/.minio minio/minio server /data
+docker run -d -p $MINIO_PORT:9000 --restart=always --name minio -e "MINIO_ACCESS_KEY=$ACCESS_KEY" -e "MINIO_SECRET_KEY=$SECRET_KEY" -v $MINIO_DATA:/data -v $MINIO_CONFIG:/root/.minio minio/minio server /data
 
 # Run minio server by service(Swarm)
 #echo $ACCESS_KEY | docker secret create access_key -
@@ -52,7 +81,7 @@ docker run -d -p 9000:9000 --restart=always --name minio -e "MINIO_ACCESS_KEY=$A
 ####################
 # Docker Registery #
 ####################
-
+REGISTRY_URL="$MASTER_ADDRESS:$REGISTRY_PORT"
 REGISTRY_CONFIG=$MLAD/master/registry
 
 mkdir -p $REGISTRY_CONFIG
@@ -65,28 +94,28 @@ log:
         service: registry
 http:
     addr: :5000
+    tls:
+        certificate: /certs/domain.crt
+        key: /certs/domain.key
 storage:
-    cache:
-        layerinfo: inmemory
     s3:
         accesskey: $ACCESS_KEY
         secretkey: $SECRET_KEY
         region: "us-east-1"
-        regionendpoint: "http://localhost:9000"
+        regionendpoint: "http://$MASTER_ADDRESS:$MINIO_PORT"
         bucket: "docker-registry"
         encrypt: false
-        secure: true
+        secure: false
         v4auth: true
         chunksize: 5242880
         rootdirectory: /
 EOL
 
 # Generate certificates to MinIO
-read -p "Private docker registry URL: " REGISTRY_URL
 REGISTRY_CERT=$MINIO_DATA/docker-registry/certs/${REGISTRY_URL/:/-}
 echo $REGISTRY_CERT
 
-if [ 0 ]; then#[ -d $REGISTRY_CERT ]; then
+if [ -d $REGISTRY_CERT ]; then
     echo "Already generated certificates."
 else
     mkdir -p $REGISTRY_CERT
@@ -106,7 +135,7 @@ fi
 # Run docker registry server
 docker stop registry > /dev/null 2>&1
 docker rm registry > /dev/null 2>&1
-docker run -d -p 5000:5000 --restart=always -e REGISTRY_HTTP_ADDR=0.0.0.0:5000 -e REGISTRY_HTTP_TLS_CERTIFICATE=/certs/domain.crt -e REGISTRY_HTTP_TLS_KEY=/certs/domain.key -v "$REGISTRY_CERT":/certs -v $REGISTRY_CONFIG:/etc/docker/registry --name registry registry:2
+docker run -d -p 5000:5000 --restart=always -v "$REGISTRY_CERT":/certs -v $REGISTRY_CONFIG:/etc/docker/registry --name registry registry:2
 
 # Run docker registry server by service(Swarm)
 # docker secret create domain.crt $REGISTRY_CONFIG/certs/domain.crt
